exercicio 8

+++ Exercicio3.js
@@ -1,7 +1,7 @@
 // Exercício 3: Ímpar ou par
 // Faça um programa que receba um número do usuário e informe se este número é par ou ímpar.
 
-let numero = parseFloat(prompt("Insira um numero qualquer: "))
+let numero = parseInt(prompt("Insira um numero qualquer: "))
 
 if (numero % 2 == 0)
 {console.log ("O numero e par")}
